// ssyweb/SSYNEW/src/app/claim/models/pf.beneficiaryBasic.model.ts
export class PFBeneficiaryDetails {
    benSno: number;
    errorMessage: string;
    benFullName: string;
    benMobileNo: string;
    benPhoto: string;
    ssI_Number: string;
    regNumber: string;
    regType: string;
    registrationPhotoCopy: null;
    benAddress: string;
    isActive: boolean;
    status: number;
    benPFStatus: number;
    benPFAccountId: number;
    benPFAccountNo: string;
    benSSINOrPFAccountNo: string;
    benOccupationId: number;
    benRegDate: Date;
    benBmc: number;
    benLastCollectionDate: Date;
}