// ssyweb/SSYNEW/src/app/models/receiptbook.model.ts
export class ReceiptBookModel {
    id: number;
    date: any;
    dateString: string;
    agentCode: number;
    bookNo: string;
    receiptNoFrom: number;
    //receiptNoFromDateString: string;
    receiptNoTo: number;
    //receiptNoToDateString: string;
    createdBy: number;
    createdDate: Date;
}